Napravljeno da se starreview moze koristitit s controller componentom
